requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
pywebpush==2.0.1
py-vapid==1.9.2
APScheduler==3.10.4
python-dotenv==1.0.1
cryptography
